Add files via upload
Made code unnecessarily more abstract and complicated to make it harder for students to put together in the 'run code' activity itself. If this was for real this just adds unneeded complexity and makes the code much more difficult to read, something which should not be done.

Also added confirmation if the user truly wants to delete all the contacts or if they want to back out of that operation.